scripts/requirements.txt
google-cloud-storage==2.17.0
pandas==2.2.2
pick==2.3.2
redis==5.0.8
rich==13.7.1
